Button: "Nullstill rute" does not clear the from [Fra] and to [Till] fields #108 - refactor
